Release 'Bit-Vector-7.4' of Bit::Vector

+++ ToolBox.h
@@ -31,8 +31,8 @@ typedef  unsigned   char    N_char;
 typedef  unsigned   char    N_byte;
 typedef  unsigned   short   N_short;
 typedef  unsigned   short   N_shortword;
-typedef  unsigned   int     N_int;
-typedef  unsigned   int     N_word;
+typedef  unsigned   long    N_int;       /* used to be "int"! */
+typedef  unsigned   long    N_word;      /* used to be "int"! */
 typedef  unsigned   long    N_long;
 typedef  unsigned   long    N_longword;
 
@@ -43,8 +43,8 @@ typedef  signed     char    Z_char;
 typedef  signed     char    Z_byte;
 typedef  signed     short   Z_short;
 typedef  signed     short   Z_shortword;
-typedef  signed     int     Z_int;
-typedef  signed     int     Z_word;
+typedef  signed     long    Z_int;       /* used to be "int"! */
+typedef  signed     long    Z_word;      /* used to be "int"! */
 typedef  signed     long    Z_long;
 typedef  signed     long    Z_longword;
 
@@ -123,11 +123,12 @@ typedef  Z_longword         *Z_longwordptr;
 /*****************************************************************************/
 
 /*****************************************************************************/
-/*  VERSION:  5.7                                                            */
+/*  VERSION:  5.8                                                            */
 /*****************************************************************************/
 /*  VERSION HISTORY:                                                         */
 /*****************************************************************************/
 /*                                                                           */
+/*    Version 5.8   03.09.13  Allows bit vectors with > 2^32 bits.           */
 /*    Version 5.7   17.05.12  Changed boolean data type.                     */
 /*    Version 5.6   27.07.09  Made it MacOS X compatible.                    */
 /*    Version 5.5   03.10.04  Added compiler directives for C++.             */
